+++ silogistik2106650185/src/main/java/apap/ti/silogistik2106650185/model/PermintaanPengiriman.java
@@ -1,5 +1,6 @@
 package apap.ti.silogistik2106650185.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Date;
 import java.util.List;
@@ -42,7 +43,7 @@ public class PermintaanPengiriman {
 
     @NotNull
     @Column(name="biaya_pengiriman", nullable = false)
-    private int biayaPengiriman;
+    private BigDecimal biayaPengiriman;
 
     @NotNull
     @Column(name="jenis_layanan", nullable = false)
@@ -62,7 +63,7 @@ public class PermintaanPengiriman {
 
     public PermintaanPengiriman(Long idPermintaanPengiriman, @NotNull @Size(max = 16, min = 16) String nomorPengiriman,
             @NotNull boolean isCancelled, @NotNull String namaPenerima, @NotNull String alamatPenerima,
-            @NotNull Date tanggalPengiriman, @NotNull int biayaPengiriman, @NotNull @Size(max = 4) int jenisLayanan,
+            @NotNull Date tanggalPengiriman, @NotNull BigDecimal biayaPengiriman, @NotNull @Size(max = 4) int jenisLayanan,
             @NotNull LocalDateTime waktuPermintaan, Karyawan karyawan,
             List<PermintaanPengirimanBarang> listPermintaanPengirimanBarang) {
         this.idPermintaanPengiriman = idPermintaanPengiriman;
@@ -129,11 +130,11 @@ public class PermintaanPengiriman {
         this.tanggalPengiriman = tanggalPengiriman;
     }
 
-    public int getBiayaPengiriman() {
+    public BigDecimal getBiayaPengiriman() {
         return biayaPengiriman;
     }
 
-    public void setBiayaPengiriman(int biayaPengiriman) {
+    public void setBiayaPengiriman(BigDecimal biayaPengiriman) {
         this.biayaPengiriman = biayaPengiriman;
     }
